# src/InvoiceReviewAssistant.Api/Controllers/InvoiceDraftsController.cs
using InvoiceReviewAssistant.Api.Contracts;
using InvoiceReviewAssistant.Core.Drafts;
using InvoiceReviewAssistant.Core.Invoices;
using Microsoft.AspNetCore.Mvc;

namespace InvoiceReviewAssistant.Api.Controllers;

[ApiController]
[Route("api/invoices/{id}/draft")]
public sealed class InvoiceDraftsController(SaveInvoiceDraftService saveService) : ControllerBase
{
    [HttpPut]
    [ProducesResponseType<InvoiceDetailDto>(StatusCodes.Status200OK)]
    [ProducesResponseType<InvoiceProblemDetails>(StatusCodes.Status400BadRequest)]
    [ProducesResponseType<InvoiceProblemDetails>(StatusCodes.Status404NotFound)]
    [ProducesResponseType<InvoiceProblemDetails>(StatusCodes.Status409Conflict)]
    [ProducesResponseType<InvoiceProblemDetails>(StatusCodes.Status500InternalServerError)]
    public async Task<IActionResult> Save(
        string id,
        SaveInvoiceDraftRequest request,
        CancellationToken cancellationToken)
    {
        if (!TryParseInvoiceId(id, out var invoiceId))
        {
            return Problem(
                StatusCodes.Status400BadRequest,
                "REQUEST_VALIDATION_FAILED",
                "The invoice identifier is invalid",
                "The invoice identifier must be a non-empty lowercase UUID.",
                new Dictionary<string, string[]>(StringComparer.Ordinal)
                {
                    ["id"] = ["The invoice identifier must be a non-empty lowercase UUID."],
                });
        }

        var result = await saveService.SaveAsync(
            invoiceId,
            request.Draft.ToDomain(),
            request.ExpectedVersion.ToDraftVersion(),
            cancellationToken);

        return result switch
        {
            SaveInvoiceDraftResult.Saved saved => Ok(saved.Invoice.ToDto(saved.Corrections)),
            SaveInvoiceDraftResult.NotFound => Problem(
                StatusCodes.Status404NotFound,
                "INVOICE_NOT_FOUND",
                "Invoice not found",
                "The requested invoice does not exist."),
            SaveInvoiceDraftResult.VersionConflict conflict => Problem(
                StatusCodes.Status409Conflict,
                "INVOICE_VERSION_CONFLICT",
                "The invoice has changed",
                "Refresh the invoice and reconcile your unsaved changes before trying again.",
                currentVersion: conflict.CurrentVersion.Value),
            SaveInvoiceDraftResult.StateConflict conflict => Problem(
                StatusCodes.Status409Conflict,
                "INVOICE_STATE_CONFLICT",
                "The invoice cannot be edited",
                "The invoice's current state does not allow draft changes.",
                currentVersion: conflict.CurrentVersion.Value),
            _ => throw new InvalidOperationException("The draft-save result is unsupported."),
        };
    }

    private ObjectResult Problem(
        int status,
        string code,
        string title,
        string detail,
        IReadOnlyDictionary<string, string[]>? fields = null,
        int? currentVersion = null)
    {
        var result = new ObjectResult(InvoiceProblems.Create(
            HttpContext,
            status,
            code,
            title,
            detail,
            fields,
            currentVersion))
        {
            StatusCode = status,
        };
        result.ContentTypes.Add("application/problem+json");
        return result;
    }

    private static bool TryParseInvoiceId(string value, out InvoiceId invoiceId)
    {
        if (value.Length == 36 &&
            string.Equals(value, value.ToLowerInvariant(), StringComparison.Ordinal) &&
            Guid.TryParseExact(value, "D", out var parsed) &&
            parsed != Guid.Empty)
        {
            invoiceId = new InvoiceId(parsed);
            return true;
        }

        invoiceId = default;
        return false;
    }
}