# src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "InvoiceReviewAssistant.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5190/"
    }
  ],
  "paths": {
    "/api/invoices/{id}/approve": {
      "post": {
        "tags": [
          "InvoiceApproval"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveInvoiceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "approveInvoice"
      }
    },
    "/api/invoices/{id}": {
      "get": {
        "tags": [
          "InvoiceDetails"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "getInvoice"
      }
    },
...
    "/api/invoices/{id}/draft": {
      "put": {
        "tags": [
          "InvoiceDrafts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveInvoiceDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "saveInvoiceDraft"
      }
    },
    "/api/invoices/{id}/export": {
      "get": {
        "tags": [
          "InvoiceExports"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceExportV1Dto"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "exportInvoice"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveInvoiceRequest": {
        "required": [
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "DocumentIntegrityStatus": {
        "type": "string",
        "enum": [
          "available",
          "missing",
          "corrupt"
        ]
      },
      "DocumentTextSource": {
        "type": "string",
        "enum": [
          "nativeText",
          "ocr"
        ]
      },
      "FieldCorrectionDto": {
        "required": [
          "id",
          "auditEventId",
          "field",
          "previousValue",
          "newValue",
          "draftVersion",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^(?:0|[1-9]\\d*)$"
          },
          "auditEventId": {
            "type": "string",
            "pattern": "^(?:0|[1-9]\\d*)$"
          },
          "field": {
            "$ref": "#/components/schemas/InvoiceFieldKey"
          },
          "previousValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "newValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "draftVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceDecisionDto": {
        "required": [
          "kind",
          "decidedAt",
          "rejectionReason"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DecisionKind"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "rejectionReason": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "InvoiceDetailDto": {
        "required": [
          "id",
          "status",
          "draftVersion",
          "lastValidatedVersion",
          "createdAt",
          "updatedAt",
          "document",
          "documentTextSource",
          "fields",
          "reviewNotes",
          "summary",
          "currentValidation",
          "corrections",
          "processingFailure",
          "decision"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "draftVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "lastValidatedVersion": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "document": {
            "$ref": "#/components/schemas/InvoiceDocumentDto"
          },
          "documentTextSource": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentTextSource"
              }
            ]
          },
          "fields": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InvoiceFieldsDto"
              }
            ]
          },
          "reviewNotes": {
            "type": [
              "null",
              "string"
            ]
          },
          "summary": {
            "$ref": "#/components/schemas/InvoiceSummaryDto"
          },
          "currentValidation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationRunDto"
              }
            ]
          },
          "corrections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldCorrectionDto"
            }
          },
          "processingFailure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProcessingFailureDto"
              }
            ]
          },
          "decision": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InvoiceDecisionDto"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "InvoiceDocumentDto": {
        "required": [
          "originalFilename",
          "mediaType",
          "byteLength",
          "sha256",
          "pageCount",
          "integrityStatus"
        ],
        "type": "object",
        "properties": {
          "originalFilename": {
            "type": "string"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "application/pdf"
            ]
          },
          "byteLength": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "integrityStatus": {
            "$ref": "#/components/schemas/DocumentIntegrityStatus"
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceExportV1Dto": {
        "required": [
          "schemaVersion",
          "invoice",
          "auditHistory"
        ],
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "enum": [
              "1.0"
            ]
          },
          "invoice": {
            "$ref": "#/components/schemas/InvoiceDetailDto"
          },
          "auditHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEventDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceFieldsDto": {
        "required": [
          "supplier",
          "reference",
          "datesAndTerms",
          "amounts"
        ],
        "type": "object",
        "properties": {
          "supplier": {
            "$ref": "#/components/schemas/SupplierFieldsDto"
          },
          "reference": {
            "$ref": "#/components/schemas/ReferenceFieldsDto"
          },
          "datesAndTerms": {
            "$ref": "#/components/schemas/DatesAndTermsFieldsDto"
          },
          "amounts": {
            "$ref": "#/components/schemas/AmountFieldsDto"
          }
        },
        "additionalProperties": false
      },
      "InvoiceProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 100
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "fields": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            ]
          },
          "currentVersion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 1
              }
            ]
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlationId",
          "fields",
          "currentVersion"
        ],
        "additionalProperties": false
      },
...
      "InvoiceStatus": {
        "type": "string",
        "enum": [
          "processing",
          "reviewRequired",
          "readyForApproval",
          "approved",
          "rejected",
          "processingFailed"
        ]
      },
      "InvoiceSummaryDto": {
        "required": [
          "extractedFieldCount",
          "warningCount",
          "errorCount",
          "manualCorrectionCount"
        ],
        "type": "object",
        "properties": {
          "extractedFieldCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "warningCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "errorCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "manualCorrectionCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "ProcessingFailureDto": {
        "required": [
          "stage",
          "code",
          "message",
          "failedAt"
        ],
        "type": "object",
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/ProcessingStage"
          },
          "code": {
            "$ref": "#/components/schemas/ProcessingFailureCode"
          },
          "message": {
            "type": "string"
          },
          "failedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          }
        },
        "additionalProperties": false
      },
      "ProcessingStage": {
        "type": "string",
        "enum": [
          "upload",
          "pdfExtraction",
          "ocr",
          "aiExtraction",
          "parsing",
          "persistence",
          "startupRecovery"
        ]
      },
...
      "SaveInvoiceDraftRequest": {
        "required": [
          "expectedVersion",
          "draft"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "draft": {
            "$ref": "#/components/schemas/InvoiceDraftInputDto"
          }
        },
        "additionalProperties": false
      },
...
      "ValidationRunDto": {
        "required": [
          "id",
          "draftVersion",
          "validatedAt",
          "warningCount",
          "errorCount",
          "results"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "draftVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "validatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "warningCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "errorCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationResultDto"
            }
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}